pyproject.toml
[project]
name = 'agents-workshop'
version = '0'
requires-python = '>=3.13,<3.14'
dependencies = [
	'pydantic>=2.8',
	'pydantic-settings>=2.4',
	'langgraph>=0.2',
	'langgraph-checkpoint-sqlite>=3.0.3',
	'langfuse>=2.50',
	'httpx',
	'structlog',
	'pyyaml',
	'fastapi',
	'uvicorn[standard]',
	'langchain-groq',
	'firecrawl-py>=1.5',
]

[dependency-groups]
dev = [
	'pytest>=8.4',
	'pytest-asyncio>=0.24',
	'ruff>=0.13',
	'ty',
]

[tool.pytest.ini_options]
pythonpath = '.'
asyncio_mode = 'auto'

[tool.ruff]
line-length = 120
[tool.ruff.format]
indent-style = 'tab'
skip-magic-trailing-comma = true
[tool.ruff.lint]
select = ['ASYNC', 'F', 'FAST', 'PLE', 'PT', 'SIM', 'TID', 'UP045']
ignore = [
	'ASYNC230',
	'SIM108',
	'SIM115',
]
[tool.ruff.lint.flake8-tidy-imports.banned-api]
'requests'.msg = 'use httpx instead'
